Create Able to vote or not.cpp
Recently I started learning C++ programming language, and here I learn how to use datatype and variable to simply write code in C++ and to practice this I wrote a program to identify whether candidate are able to vote or not. (Condition is if candidate age is equal to or more than 18 then they are able to vote else not able to vote).